Whole project got commented, css files are seperated and commented for dahsboard elements, election controlers are seperated in individual js files
