+++ src/styles/dashboard-add-election.css
@@ -1,0 +1,110 @@
+/* Add Election Form */
+.add-election-form input {
+  width: 25vw;
+}
+
+.add-election-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+/* Add Election Modal */
+.add-election {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(1, 1, 1, 0.2);
+  backdrop-filter: blur(5px);
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+/* Add Election Form */
+.options,
+.topic,
+.election-name,
+.election-form {
+  margin: 0.5rem;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+/* Add Option Button */
+
+.button-box {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+}
+
+.add-option-btn {
+  align-items: center;
+  justify-content: center;
+  padding: 1rem;
+  margin: 0.5rem;
+  border: none;
+  background-color: transparent;
+  width: 9rem;
+  height: 4rem;
+  border-radius: 0.5rem;
+  transition: 0.3s;
+  border: 0.2px solid #bdc3c7;
+}
+
+.next-btn,
+.submit-btn {
+  align-items: center;
+  justify-content: center;
+  padding: 2.3rem;
+}
+
+.next-btn,
+.next-img,
+.submit-btn,
+.submit-img {
+  width: 4rem;
+  height: 3rem;
+  margin: 0.3rem;
+  border: none;
+  border-radius: 1rem;
+  transition: 0.3s;
+}
+
+.submit-btn:hover {
+  background-color: #27ae60;
+}
+
+/* Added Info */
+.added-info {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  margin: 0.2rem;
+  font-size: 1.2rem;
+  color: #2ecc71;
+  border: 0.5px solid black;
+  border-radius: 2rem;
+  padding: 1rem;
+  outline: none; /* Remove focus outline */
+  user-select: none; /* Prevent selection */
+}
+
+@media (max-width: 900px) {
+  /* Add Election Form */
+  .add-election-form input {
+    max-width: none;
+    width: 55vw;
+  }
+  .add-election {
+    padding: 0px;
+    margin: 0px;
+  }
+}
