+++ src/styles/dashboard-home.css
@@ -1,0 +1,194 @@
+/* One Election Form */
+
+.cont-one-el {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: start;
+  background-color: #ecf0f1;
+  max-width: 70vw;
+  border-radius: 2rem;
+  padding: 3rem;
+  max-height: 85vh;
+  transition: 0.3s;
+  overflow: auto;
+}
+
+.one-election {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(1, 1, 1, 0.2);
+  backdrop-filter: blur(5px);
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.one-el-form input {
+  width: 25vw;
+}
+
+.one-el-title {
+  text-align: center;
+  padding: 1rem;
+}
+
+.one-el-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.close-btn-one-el {
+  padding: 0px;
+  margin: 1rem;
+  margin-right: -94%;
+  margin-top: 0%;
+  border: none;
+  width: 2.5rem;
+  height: 2.5rem;
+  border-radius: 50em;
+}
+
+.el-name {
+  width: 10rem;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
+
+/* Election edit buttons */
+.add-topic,
+.delete-el-btn,
+.update-el {
+  font-size: 1rem;
+  color: #ecf0f1;
+  width: auto;
+  padding: 1.2rem;
+  border-radius: 3rem;
+  border: none;
+  margin: 0.2rem;
+  transition: 0.2s;
+}
+
+.add-option-btn-update {
+  align-items: center;
+  justify-content: center;
+  padding: 1rem;
+  margin: 0.5rem;
+  border: none;
+  background-color: transparent;
+  width: 9rem;
+  height: 4rem;
+  border-radius: 0.5rem;
+  transition: 0.3s;
+  border: 0.2px solid #bdc3c7;
+}
+
+.add-topic {
+  background-color: #f39c12;
+}
+
+.delete-el-btn {
+  background-color: #d35400;
+}
+
+.update-el {
+  background-color: #27ae60;
+}
+
+.add-topic:hover,
+.delete-el-btn:hover,
+.update-el:hover {
+  opacity: 0.5;
+}
+
+/* Warning */
+.warning {
+  text-align: center;
+  color: #b00020;
+  margin: 1rem;
+}
+
+/* Election Button view */
+.election-btn {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  margin: 0.3rem;
+  width: auto;
+  height: auto;
+  padding: 1rem;
+  border-radius: 1rem;
+  transition: 0.3s;
+  border: none;
+  color: #34495e;
+  box-shadow: rgb(204, 219, 232) 3px 3px 6px 0px inset,
+    rgba(255, 255, 255, 0.5) -3px -3px 6px 1px inset;
+  background-color: #d7e2ec;
+  border-bottom: 1px solid #bdc3c7;
+  border-right: 1px solid #bdc3c7;
+}
+
+.election-btn:hover {
+  background-color: #e5f1f3;
+}
+
+/* Election Image view */
+.election-img {
+  margin: 0.5rem;
+  width: 3.5vw;
+  height: auto;
+  padding: 0.2rem;
+}
+
+/* Status Indicators view*/
+.inProgress {
+  opacity: 0.8;
+  border-right: 1px solid #d35400;
+  border-bottom: 1px solid #d35400;
+}
+
+.done {
+  opacity: 0.7;
+  border-right: 1px solid #27ae60;
+  border-bottom: 1px solid #27ae60;
+}
+
+@media (max-width: 900px) {
+  /* Election Button view */
+  .election-btn {
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    width: 75vw;
+  }
+
+  /* One Election Form */
+  .one-el-form input {
+    max-width: none;
+    width: 55vw;
+  }
+
+  .cont-one-el {
+    padding: 2rem;
+    margin: 0px;
+    min-height: 100vh;
+    min-width: 100vw;
+  }
+
+  .one-el-form {
+    margin: 0px;
+    width: 100%;
+  }
+
+  .one-el-btns {
+    display: flex;
+    flex-direction: column;
+  }
+}
