+++ src/styles/dashboard-results.css
@@ -1,0 +1,49 @@
+/* Chart Styles */
+#charts {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2rem;
+  justify-content: center;
+  margin-bottom: 3rem;
+}
+
+.chart-container {
+  min-width: 20vw;
+  padding: 1rem;
+  border-radius: 12px;
+  background-color: #f9f9f9;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.chart-container h4 {
+  text-align: center;
+  margin-bottom: 12px;
+  font-size: 16px;
+}
+
+.el-res {
+  font-size: 2rem;
+  margin: 2rem;
+}
+
+@media (max-width: 900px) {
+  .chart-container {
+    min-width: 0px;
+    padding: 0.5rem;
+    width: 75vw;
+    border-radius: 12px;
+    background-color: #f9f9f9;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .chart-container h4 {
+    text-align: center;
+    margin-bottom: 12px;
+    font-size: 0.8rem;
+  }
+}
