+++ src/styles/dashboard-users.css
@@ -1,0 +1,290 @@
+/* User Management Styles */
+
+.user-content input {
+  max-width: 16vw;
+  min-width: 13vw;
+  width: auto;
+}
+
+.see-user-box {
+  flex-direction: column;
+}
+
+.see-user-box {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(1, 1, 1, 0.2);
+  backdrop-filter: blur(5px);
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.titles {
+  background-color: #ecf0f1;
+  padding: 1rem;
+  border-radius: 2rem;
+  margin: 0.5rem;
+  border: 1px solid #34495e;
+  outline: none; /* Remove focus outline */
+  user-select: none; /* Prevent selection */
+}
+
+.user-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: #ecf0f1;
+  max-width: 70vw;
+  border-radius: 2rem;
+  padding: 3rem;
+  margin: 1rem;
+  max-height: none;
+  max-width: 100vw;
+  transition: 0.3s;
+  overflow: auto;
+}
+
+.create-user-cont input {
+  width: 25vw;
+}
+
+.create-user-cont {
+  display: flex;
+  background-color: #ecf0f1;
+  align-items: center;
+  justify-content: center;
+  border-radius: 1rem;
+  width: 80vw;
+  height: fit-content;
+  padding: 2rem;
+  overflow: hidden;
+  border: 1px solid #7f8c8d;
+}
+
+.create-user-cont button {
+  margin: 0.3rem;
+  height: auto;
+  align-items: center;
+  justify-content: center;
+  padding: 0.5rem;
+}
+
+.users-list {
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  align-items: start;
+  margin-top: 2rem;
+  width: fit-content;
+  min-height: 14rem;
+  overflow: auto;
+}
+
+.one-user {
+  display: flex;
+  align-items: center;
+  justify-content: start;
+  max-width: 100vw;
+  height: auto;
+  padding: 1.5rem;
+  margin: 1rem;
+  overflow: hidden;
+}
+
+.one-user :first-child {
+  margin-right: 0.5rem;
+}
+
+#MyAccount {
+  font-size: 1rem;
+  border: none;
+  text-align: center;
+  margin: 0.2rem;
+  padding: 0.2rem 0.2rem;
+}
+
+/* User login status */
+.last-user-login {
+  text-align: center;
+}
+
+.never-logged {
+  color: #010c16;
+  background-color: rgba(255, 166, 0, 0.641);
+}
+
+.logged-six-month {
+  color: #010c16;
+  background-color: #27ae5fc1;
+}
+
+.logged-old {
+  background-color: #ee073195;
+  color: #010c16;
+}
+
+/* User section and role */
+.section,
+.role {
+  align-items: center;
+  justify-content: center;
+  padding: 0.4rem;
+  margin: 0.1rem;
+  border: 1px solid black;
+  border-radius: 0.4rem;
+  font-size: 1rem;
+}
+
+.reset-password,
+.cancel,
+.save-user,
+.delete-user,
+.edit-user {
+  padding: 0.7rem;
+  border-radius: 0.5rem;
+  border: none;
+  margin: 0.2rem;
+  transition: 0.2s;
+}
+
+.create-user-btn {
+  background-color: #27ae60;
+  color: #ecf0f1;
+  box-shadow: 0 0 3px rgba(0, 0, 0, 0.392);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0.5rem;
+  margin: 0.5rem;
+  border-radius: 0.5rem;
+  border: none;
+  transition: 0.2s;
+}
+
+.create-user-btn:hover {
+  background-color: #27ae5fa0;
+  color: #224c77;
+}
+
+.save-user {
+  background-color: #27ae60;
+  color: #f7fafa;
+}
+
+.delete-user {
+  margin-left: 0.3rem;
+  margin-right: 0.3rem;
+  background-color: #d35400;
+  color: #f7fafa;
+}
+
+.cancel,
+.edit-user {
+  background-color: #f4d03f;
+  color: #34495e;
+}
+
+.reset-password {
+  background-color: #224c77;
+  color: #f2f6f9;
+}
+
+.reset-password:hover,
+.cancel:hover,
+.save-user:hover,
+.delete-user:hover,
+.edit-user:hover {
+  opacity: 0.5;
+}
+@media (max-width: 900px) {
+  .titles {
+    font-size: 0.7rem;
+    padding: 0.5rem;
+    margin: 0.3rem;
+  }
+
+  .create-user-cont input {
+    max-width: none;
+    width: 55vw;
+  }
+
+  .user-content {
+    width: 100%;
+    padding: 0px;
+    margin: 0;
+    margin-bottom: 0px;
+    overflow: hidden;
+  }
+
+  .create-user-cont {
+    min-height: fit-content;
+    width: 100%;
+    flex-direction: column;
+    overflow: auto;
+    height: fit-content;
+    padding: 0.5rem 0rem;
+    margin: 0.5rem 0rem;
+  }
+
+  .users-list {
+    border: 1px solid gainsboro;
+    align-items: center;
+    border-radius: 1rem;
+    padding: 1rem;
+    margin: 0px;
+    width: 100%;
+    height: 100%;
+    overflow: scroll;
+    overflow-x: hidden;
+  }
+
+  .one-user {
+    display: flex;
+    align-items: center;
+    padding: 1rem;
+    min-height: fit-content;
+    border: 2px solid rgba(128, 128, 128, 0.249);
+    border-radius: 1rem;
+    width: 65vw;
+    flex-direction: column;
+  }
+
+  .one-user :first-child {
+    margin-right: 0rem;
+    margin-bottom: 0.5rem;
+  }
+
+  .one-user button {
+    margin: 0.3rem;
+    height: auto;
+    align-items: center;
+    justify-content: center;
+    padding: 0.5rem;
+  }
+
+  .create-user-btn {
+    width: auto;
+  }
+
+  .section,
+  .role {
+    width: auto;
+  }
+
+  .cancel,
+  .delete-user,
+  .save-user,
+  .edit-user {
+    width: 6rem;
+  }
+
+  .user-content input {
+    max-width: none;
+    width: 55vw;
+  }
+}
