+++ src/styles/message.css
@@ -1,3 +1,4 @@
+/* Message Container */
 .message {
   flex-direction: column;
   align-items: center;
