+++ src/styles/reset-password.css
@@ -1,3 +1,4 @@
+/* Input Fields */
 input {
   width: 20rem;
   padding: 1rem;
@@ -9,6 +10,7 @@ input {
   border-color: #34495e;
 }
 
+/* Reset Password Form */
 .reset-form {
   display: flex;
   flex-wrap: wrap;
@@ -29,6 +31,7 @@ input {
   justify-content: center;
 }
 
+/* Title */
 .title {
   text-align: center;
   color: #34495e;
@@ -42,6 +45,7 @@ input {
   border-top: 1px solid #7f8c8d;
 }
 
+/* Reset Button */
 .reset {
   align-items: center;
   appearance: none;
