+++ src/styles/user-voter.css
@@ -1,3 +1,4 @@
+/* General styles */
 #section,
 input {
   width: 15rem;
@@ -6,7 +7,7 @@ input {
   text-align: center;
   border-radius: 2rem;
 }
-
+/* User Box general styles */
 .user-box {
   position: fixed;
   top: 0;
@@ -42,6 +43,7 @@ input {
   margin-bottom: 1rem;
 }
 
+/* Close Button */
 .close-btn {
   padding: 0px;
   margin: 1rem;
@@ -54,6 +56,16 @@ input {
   transition: 0.3s;
 }
 
+/* Buttons Box */
+
+.buttons-box {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin: 0.7rem;
+}
+
+/* cancel, save, edit and delete buttons general styles */
 .cancel,
 .save,
 .edit-account,
@@ -68,32 +80,28 @@ input {
   transition: 0.3s;
 }
 
-.buttons-box {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0.7rem;
-}
-
-.save {
-  background-color: #27ae60;
-}
-
-.delete-user {
-  background-color: #d35400;
-}
-
+/* Cancel, save, delete and Edit Account buttons individual styles */
 .cancel,
 .edit-account {
   background-color: #f4d03f;
   color: #34495e;
 }
 
+.delete-user {
+  background-color: #d35400;
+}
+
+.save {
+  background-color: #27ae60;
+}
+
+/* Edit Account and Delete User buttons hover effects */
 .edit-account:hover,
 .delete-user:hover {
   opacity: 0.6;
 }
 
+/* Close Button */
 .close-btn:hover {
   opacity: 0.4;
 }
@@ -102,6 +110,7 @@ input {
   width: 2.5rem;
 }
 
+/* Responsive Design */
 @media (max-width: 900px) {
   .buttons-box {
     flex-direction: column;
