+++ src/styles/voter.css
@@ -1,3 +1,4 @@
+/* Header */
 header {
   display: flex;
   align-items: center;
@@ -10,6 +11,25 @@ header {
   box-shadow: 0 0 4px rgba(0, 0, 0, 0.392);
   border-radius: 3rem;
   color: #7f8c8d;
+}
+
+/* Main Content */
+main {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 0.5rem auto;
+  width: 90%;
+  flex-grow: 1;
+}
+
+.tok-manual input {
+  padding: 0.8rem;
+  border-radius: 1rem;
+  max-width: 20vw;
+  min-width: 16vw;
+  width: auto;
 }
 
 .control-btn {
@@ -54,16 +74,6 @@ header {
   font-weight: 400;
 }
 
-main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 0.5rem auto;
-  width: 90%;
-  flex-grow: 1;
-}
-
 .verify-box {
   align-items: center;
   justify-content: center;
@@ -94,6 +104,7 @@ main {
   pointer-events: none;
 }
 
+/* Buttons */
 .tok-manual button,
 .no-qr-btn,
 .vote-submit-box button,
@@ -166,6 +177,7 @@ main {
   transform: translateY(2px);
 }
 
+/* Error Message */
 .error-message {
   display: flex;
   flex-direction: column;
@@ -176,6 +188,7 @@ main {
   padding: 1rem;
 }
 
+/* Question Box */
 .question-box {
   display: flex;
   flex-direction: column;
@@ -184,6 +197,7 @@ main {
   padding: 0.5rem;
 }
 
+/* Terms and Thanks Message */
 .terms,
 .thanks-message {
   text-align: center;
@@ -194,6 +208,7 @@ main {
   padding: 0.5rem;
 }
 
+/* Vote Box */
 .vote-box {
   flex-direction: column;
   max-width: 70vw;
@@ -208,6 +223,7 @@ main {
     rgba(0, 0, 0, 0.3) 0px 18px 36px -18px inset;
 }
 
+/* Vote Submit Box */
 .vote-submit-box {
   display: flex;
   flex-direction: column;
@@ -215,18 +231,24 @@ main {
   justify-content: center;
 }
 
+/* Question Text */
+.question-text {
+  text-align: center;
+}
+
+/* Current Element */
+#current-el {
+  font-size: 0.9rem;
+  margin: 1.5rem;
+  text-align: center;
+  padding: 1rem;
+}
+
+/* Radio Button Wrapper */
 .radio-wrapper-19 input[type="radio"]:focus {
   outline: 0;
   border-color: #7f8c8d;
   box-shadow: 0 0 0 4px #bdc3c7;
-}
-
-.tok-manual input {
-  padding: 0.8rem;
-  border-radius: 1rem;
-  max-width: 20vw;
-  min-width: 16vw;
-  width: auto;
 }
 
 .radio-wrapper-19 {
@@ -274,17 +296,6 @@ main {
 
 .radio-wrapper-19 label:last-child span {
   border-radius: 0 0.375em 0.375em 0;
-}
-
-.question-text {
-  text-align: center;
-}
-
-#current-el {
-  font-size: 0.9rem;
-  margin: 1.5rem;
-  text-align: center;
-  padding: 1rem;
 }
 
 @media (max-width: 900px) {
@@ -352,4 +363,3 @@ main {
     width: 80%;
   }
 }
-
