+++ src/utils/message.js
@@ -1,24 +1,25 @@
+// Description: This module provides a function to display messages in the UI, either as an error or success message.
 export function message(text, style, time) {
-  const messageBox = document.querySelector('.message')
-  messageBox.innerHTML = text
+  const messageBox = document.querySelector(".message");
+  messageBox.innerHTML = text;
   const errorDiv = document.querySelector(".message");
   errorDiv.classList.remove("hidden");
   errorDiv.classList.add("show");
 
-  if(style === 'OK'){
-    errorDiv.classList.remove('error')
-    errorDiv.classList.add('OK')
+  if (style === "OK") {
+    errorDiv.classList.remove("error");
+    errorDiv.classList.add("OK");
   } else {
-    errorDiv.classList.remove("OK")
-    errorDiv.classList.add("error")
+    errorDiv.classList.remove("OK");
+    errorDiv.classList.add("error");
   }
 
-  if(!time){
-    time = 5000; 
+  if (!time) {
+    time = 5000;
   }
 
   setTimeout(() => {
-    errorDiv.classList.remove('show')
-    errorDiv.classList.add('hidden')
+    errorDiv.classList.remove("show");
+    errorDiv.classList.add("hidden");
   }, time);
 }
